+++ example.config.js
@@ -15,6 +15,7 @@ const config = {
 	notification: {
 		enabled: false, // Set to true to enable notification
 		skipCheck: false, // Set to true if you want to keep sending notification even if stamina is above the threshold and capped.
+		codes: false, // Set to true to enable code notification (THIS ONLY CHECK FOR NEW GLOBAL CODES)
 		service: {
 			discord: {
 				enabled: false, // Set to true to enable Discord notification